+++ src/aks-preview/HISTORY.md
@@ -2,6 +2,12 @@
 
 Release History
 ===============
+0.4.11
++++++
+* Add support for load-balancer-profile
+* Set default availability type to VMSS
+* Set default load balancer SKU to Standard
+
 0.4.10
 +++++
 * Add support for `az aks update --disable-acr --acr <name-or-id>`